FEBioStudio: Removed duplicate CColorMap class.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBioStudio@12470 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ MeshTools/GDiscreteObject.cpp
@@ -193,7 +193,7 @@ void GDiscreteElementSet::AddElement(int n0, int n1)
 
 	// set default name
 	char szbuf[32] = {0};
-	sprintf(szbuf, "spring%d", m_elem.size()+1);
+	sprintf(szbuf, "spring%d", (int)m_elem.size()+1);
 	el->SetName(szbuf);
 
 	// add it to the pile
